refactor: move month/year-counter states to useDaysCalendar hook
